\name{huex10stv274probe}
\docType{data}
\alias{huex10stv274probe}
\title{Probe sequence for Affymetrix Exon microarrays.}
\description{This data object contains probe sequences for all probes/probesets defined in the huex10stv274cdf CDF file/package version 0.3.0.}
\usage{data(huex10stv274probe)}
\format{A data frame with xx rows and 6 columns, as follows.
  \tabular{lll}{
    sequence \tab character \tab probe sequence \cr
    x  \tab integer  \tab x-coordinate on the array \cr
    y  \tab integer  \tab y-coordinate on the array \cr
    Probe.Set.Name      \tab character \tab Affymetrix Probe Set Name \cr
    Probe.Interrogation.Position \tab integer \tab Probe Interrogation Position \cr
    Target.Strandedness \tab factor \tab Target Strandedness\cr
}}

\source{The probe sequence data was provided by Affymetrix. The probe file was generated by the make_cdf_from_db.pl perl function for the generation of custom/remapped CDF files for Affymetrix Exon arrays.}

\keyword{datasets}
\examples{
  data( huex10stv274probe )
  as.data.frame(huex10stv274probe[1:3,])
}
